add handler cteate_game via tg

+++ app/infrastructure/store/sqlalchemy/models/chat/chat.py
@@ -1,8 +1,8 @@
-from sqlalchemy import Column, String, Integer
+from sqlalchemy import Column, String, BigInteger
 from sqlalchemy.dialects.postgresql import UUID
 import uuid
 
-from app.infrastructure.store.sqlalchemy.model import Base
+from app.infrastructure.store.sqlalchemy.models import Base
 
 from app.core.chat import entities
 
@@ -13,7 +13,7 @@ class Chat(Base):
     id = Column(
         UUID(as_uuid=True), primary_key=True, default=uuid.uuid4
     )
-    tg_id = Column(Integer, nullable=False, unique=True)
+    tg_id = Column(BigInteger, nullable=False, unique=True)
     name = Column(String, nullable=False)
